requirements.txt
transformers
sentence-transformers
faiss-cpu
gradio
torch
